[ADD] hitter animation, distance ui [FIX] pci ui and hitting system

+++ Assets/Scripts/v2/GameManager.cs
@@ -37,7 +37,7 @@ public class GameManager : MonoBehaviour
         }
     }
 
-    public void UpdateDistance(int newDistance)
+    public void UpdateDistance(float newDistance)
     {        
         UIManager.instance.UpdateDistanceText(newDistance);
     }
